requirements.txt
fastapi==0.95.0
uvicorn==0.22.0
celery==5.2.7
redis==4.5.1
SQLAlchemy==1.4.46
requests==2.28.2
ffmpeg-python==0.2.0
python-dotenv==1.0.0  # optionally, for environment variable management
